Restructure expiry code
